requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.5
python-dotenv==0.13.0
sendgrid==6.3.0
uWSGI==2.0.18
